feat: apply changed service main design
