+++ apps/web/src/services/alert-manager/components/ServiceListContent.vue
@@ -1,0 +1,357 @@
+<script setup lang="ts">
+import { computed, reactive } from 'vue';
+import type { TranslateResult } from 'vue-i18n';
+import { useRouter } from 'vue-router/composables';
+
+import {
+    PSelectCard, PI, PLazyImg, PDivider, PTextButton,
+} from '@cloudforet/mirinae';
+
+import type { EscalationPolicyModel } from '@/schema/alert-manager/escalation-policy/model';
+import type { ServiceModel } from '@/schema/alert-manager/service/model';
+import { i18n } from '@/translations';
+
+import { useAllReferenceStore } from '@/store/reference/all-reference-store';
+import type { PluginReferenceMap } from '@/store/reference/plugin-reference-store';
+import type { WebhookReferenceMap } from '@/store/reference/webhook-reference-store';
+
+import { useProperRouteLocation } from '@/common/composables/proper-route-location';
+
+import { gray, green } from '@/styles/colors';
+
+import { SERVICE_DETAIL_TABS } from '@/services/alert-manager/constants/common-constant';
+import { ALERT_MANAGER_ROUTE } from '@/services/alert-manager/routes/route-constant';
+import { useServiceDetailPageStore } from '@/services/alert-manager/stores/service-detail-page-store';
+
+interface Props {
+    list: ServiceModel[];
+    escalationPolicyList: EscalationPolicyModel[];
+    type: 'alert' | 'healthy';
+}
+
+const props = withDefaults(defineProps<Props>(), {
+    list: undefined,
+    escalationPolicyList: undefined,
+    type: undefined,
+});
+
+const allReferenceStore = useAllReferenceStore();
+const allReferenceGetters = allReferenceStore.getters;
+const serviceDetailPageStore = useServiceDetailPageStore();
+
+const { getProperRouteLocation } = useProperRouteLocation();
+
+const router = useRouter();
+
+const storeState = reactive({
+    plugins: computed<PluginReferenceMap>(() => allReferenceGetters.plugin),
+    webhook: computed<WebhookReferenceMap>(() => allReferenceGetters.webhook),
+});
+const state = reactive({
+    isCollapsed: false,
+    title: computed<TranslateResult>(() => {
+        if (props.type === 'alert') {
+            return i18n.t('ALERT_MANAGER.SERVICE.OPEN_ALERTS_SERVICE');
+        }
+        return i18n.t('ALERT_MANAGER.SERVICE.HEALTHY_SERVICE');
+    }),
+});
+
+const getWebhookIcon = (id: string): string|undefined => {
+    const webhook = storeState.webhook[id].data;
+    if (!webhook) return undefined;
+    return storeState.plugins[webhook.plugin_info.plugin_id]?.icon || '';
+};
+const getEscalationPolicyName = (id: string): string => {
+    const escalationPolicy = props.escalationPolicyList.find((item) => item.escalation_policy_id === id);
+    return escalationPolicy?.name || '';
+};
+
+const handleClickCollapsibleTitle = () => {
+    state.isCollapsed = !state.isCollapsed;
+};
+const handleClickServiceItem = (id: string) => {
+    router.push(getProperRouteLocation({
+        name: ALERT_MANAGER_ROUTE.SERVICE.DETAIL._NAME,
+        params: {
+            serviceId: id,
+        },
+        query: {
+            tab: SERVICE_DETAIL_TABS.OVERVIEW,
+        },
+    }));
+};
+const handleClickWebhookItem = (id: string, webhookId?: string) => {
+    if (webhookId) {
+        serviceDetailPageStore.setSelectedWebhookId(webhookId);
+    }
+    router.push(getProperRouteLocation({
+        name: ALERT_MANAGER_ROUTE.SERVICE.DETAIL._NAME,
+        params: {
+            serviceId: id,
+        },
+        query: {
+            tab: SERVICE_DETAIL_TABS.WEBHOOK,
+        },
+    }));
+};
+const handleClickEscalationPolicy = (id: string, escalationPolicyId: string) => {
+    if (id) {
+        serviceDetailPageStore.setSelectedEscalationPolicyId(escalationPolicyId);
+    }
+    router.push(getProperRouteLocation({
+        name: ALERT_MANAGER_ROUTE.SERVICE.DETAIL._NAME,
+        params: {
+            serviceId: id,
+        },
+        query: {
+            tab: SERVICE_DETAIL_TABS.SETTINGS,
+        },
+    }));
+};
+</script>
+
+<template>
+    <div class="service-list-content"
+         :class="{ 'is-collapsed': state.isCollapsed }"
+    >
+        <div class="collapsible-title flex items-center font-medium"
+             @click="handleClickCollapsibleTitle"
+        >
+            <p-i name="ic_chevron-down"
+                 width="1.25rem"
+                 height="1.25rem"
+                 :color="gray[900]"
+                 class="arrow-button"
+            />
+            <span>{{ state.title }}</span>
+        </div>
+        <div class="collapsible-contents flex flex-wrap gap-4">
+            <p-select-card v-for="(item, idx) in props.list"
+                           :key="`service-item-${idx}`"
+                           class="card"
+                           @change="handleClickServiceItem(item.service_id)"
+            >
+                <div class="card-inner-wrapper">
+                    <div class="flex items-center justify-between">
+                        <p class="flex-1 text-label-xl font-bold truncate">
+                            {{ item.name }}
+                        </p>
+                        <p-i name="ic_chevron-right"
+                             width="1.5rem"
+                             height="1.5rem"
+                             :color="gray[500]"
+                        />
+                    </div>
+                    <div class="contents">
+                        <div class="inner">
+                            <div v-if="props.type === 'alert'"
+                                 class="flex flex-col gap-3"
+                            >
+                                <div class="alerts">
+                                    <p class="title text-gray-700">
+                                        {{ $t('ALERT_MANAGER.SERVICE.OPEN_ALERTS') }}
+                                    </p>
+                                    <p class="text-display-md font-medium">
+                                        {{ (item?.alerts.TRIGGERED?.HIGH || 0) + (item?.alerts.TRIGGERED?.LOW || 0)
+                                            + (item?.alerts.ACKNOWLEDGED?.HIGH || 0) + (item?.alerts.ACKNOWLEDGED?.LOW || 0) }}
+                                    </p>
+                                </div>
+                                <div class="flex justify-between pl-2 pr-2">
+                                    <div class="alerts triggered text-red-500">
+                                        <p class="title">
+                                            {{ $t('ALERT_MANAGER.ALERTS.TRIGGERED') }}
+                                        </p>
+                                        <div class="triggered-info">
+                                            <p class="count">
+                                                {{ (item?.alerts.TRIGGERED?.HIGH || 0) + (item?.alerts.TRIGGERED?.LOW || 0) }}
+                                            </p>
+                                            <div class="ml-2 text-label-sm text-gray-700">
+                                                <span class="pl-1">{{ $t('ALERT_MANAGER.ALERTS.HIGH') }}:</span>
+                                                <span> {{ item?.alerts.TRIGGERED?.HIGH || 0 }}</span>
+                                            </div>
+                                        </div>
+                                    </div>
+                                    <div class="alerts acknowledged text-gray-600">
+                                        <p class="title">
+                                            {{ $t('ALERT_MANAGER.ALERTS.ACKNOWLEDGED') }}
+                                        </p>
+                                        <div class="triggered-info">
+                                            <p class="count">
+                                                {{ (item?.alerts.ACKNOWLEDGED?.HIGH || 0) + (item?.alerts.ACKNOWLEDGED?.LOW || 0) }}
+                                            </p>
+                                        </div>
+                                    </div>
+                                </div>
+                            </div>
+                            <div v-else
+                                 class="flex pt-3 pb-3"
+                            >
+                                <div class="flex items-center gap-2 text-green-600">
+                                    <p-i name="ic_check_healthy"
+                                         width="1.5rem"
+                                         height="1.5rem"
+                                         :color="green[600]"
+                                    />
+                                    <p class="title">
+                                        {{ $t('ALERT_MANAGER.ALERTS.HEALTHY') }}
+                                    </p>
+                                </div>
+                            </div>
+                            <p-divider class="block mt-1 mb-1" />
+                            <div class="additional-info-wrapper">
+                                <div>
+                                    <p class="title">
+                                        {{ $t('ALERT_MANAGER.SERVICE.WEBHOOK', { cnt: item?.webhooks?.length || 0 }) }}
+                                    </p>
+                                    <div class="flex items-center">
+                                        <span v-for="(webhook, webhookIdx) in item.webhooks?.slice(0,5)"
+                                              :key="`webhook-item-${webhookIdx}`"
+                                              class="webhook"
+                                              @click.stop="handleClickWebhookItem(item.service_id, webhook)"
+                                        >
+                                            <p-lazy-img :src="getWebhookIcon(webhook)"
+                                                        error-icon="ic_webhook"
+                                                        width="0.875rem"
+                                                        height="0.875rem"
+                                                        class="icon"
+                                            />
+                                        </span>
+                                        <span class="webhook chevron"
+                                              @click.stop="handleClickWebhookItem(item.service_id)"
+                                        >
+                                            <p-i
+                                                name="ic_chevron-right"
+                                                width="1.125rem"
+                                                height="1.125rem"
+                                                color="inherit"
+                                            />
+                                        </span>
+                                    </div>
+                                </div>
+                                <div class="flex flex-col gap-1 items-end">
+                                    <p class="title">
+                                        {{ $t('ALERT_MANAGER.ESCALATION_POLICY.TITLE', { cnt: 11 }) }}
+                                    </p>
+                                    <p-text-button @click.stop="handleClickEscalationPolicy(item.service_id, item.escalation_policy_id)">
+                                        <p class="truncate text-blue-700 pr-1 pl-1">
+                                            {{ getEscalationPolicyName(item.escalation_policy_id) }}
+                                        </p>
+                                    </p-text-button>
+                                </div>
+                            </div>
+                        </div>
+                    </div>
+                </div>
+            </p-select-card>
+        </div>
+    </div>
+</template>
+
+<style scoped lang="postcss">
+.service-list-content {
+    .collapsible-title {
+        height: 1.5rem;
+        cursor: pointer;
+        margin-bottom: 0.5rem;
+        .arrow-button {
+            transition: transform 0.3s ease-in-out;
+        }
+    }
+    .collapsible-contents {
+        opacity: 1;
+        transition: opacity 0.3s ease, visibility 0.3s ease;
+    }
+    &.is-collapsed {
+        .collapsible-title {
+            .arrow-button {
+                transform: rotate(-90deg);
+            }
+        }
+        .collapsible-contents {
+            display: none;
+            height: 0;
+            margin: 0;
+            padding: 0;
+            opacity: 0;
+            transition: opacity 0s ease;
+        }
+    }
+    .card {
+        width: 28rem;
+        max-width: 28rem;
+        padding: 1.25rem 1.5rem 1rem 1.5rem;
+        box-shadow: 0 2px 4px 0 rgba(0, 0, 0, 0.06);
+        .card-inner-wrapper {
+            @apply flex flex-col w-full;
+            gap: 1.25rem;
+            .contents {
+                @apply flex gap-8;
+                .inner {
+                    @apply flex flex-col flex-1 gap-3;
+                }
+                .alerts {
+                    @apply relative flex flex-col;
+                    gap: 0.25rem;
+                    width: calc(50% - 1rem);
+                    .title {
+                        @apply text-label-md font-medium;
+                    }
+                    .count {
+                        @apply text-display-sm;
+                    }
+                    .triggered-info {
+                        @apply flex items-center;
+                    }
+                    &.triggered {
+                        padding-left: 1rem;
+                        &::before {
+                            @apply absolute bg-red-400;
+                            content: '';
+                            width: 0.063rem;
+                            height: 100%;
+                            top: 0;
+                            left: 0;
+                        }
+                    }
+                    &.acknowledged {
+                        padding-left: 1rem;
+                        &::before {
+                            @apply absolute bg-gray-300;
+                            content: '';
+                            width: 0.063rem;
+                            height: 100%;
+                            top: 0;
+                            left: 0;
+                        }
+                    }
+                }
+                .additional-info-wrapper {
+                    @apply flex justify-between;
+                    flex: 1;
+                    .title {
+                        @apply text-paragraph-md text-gray-600;
+                    }
+                    .webhook {
+                        @apply flex items-center justify-center rounded-full bg-gray-100 border border-white ;
+                        width: 1.5rem;
+                        height: 1.5rem;
+                        &:hover {
+                            @apply bg-blue-200;
+                        }
+                        & + .webhook {
+                            margin-left: -0.25rem;
+                        }
+                        &.chevron {
+                            @apply border-gray-200;
+                        }
+                        .icon {
+                            margin-bottom: 0;
+                        }
+                    }
+                }
+            }
+        }
+    }
+}
+</style>
